You should be able to get previously loaned books

+++ library/schema.sql
@@ -43,9 +43,9 @@ create table loans (
     loan_id INTEGER primary key autoincrement,
     book_id INTEGER not null,
     user_id INTEGER not null,
-    loan_date INTEGER not null,
-    return_date INTEGER,
-    due_date INTEGER,
+    loan_date REAL not null,
+    return_date REAL,
+    due_date REAL,
     FOREIGN KEY(book_id) REFERENCES books(book_id)
 );
